+++ components/overlays/NoteDetailsOverlay.vue
@@ -27,12 +27,12 @@
 </template>
 
 <script setup lang="ts">
-import OverlayDialog from "@/components/ui/OverlayDialog.vue";
+import OverlayDialog from "~/components/overlays/OverlayDialog.vue";
 import { defineProps } from "vue";
-import { Note, NoteDirection } from "@/lib/notes";
-import { formatDate } from "@/helpers/date";
+import { Note, NoteDirection } from "~/lib/notes";
+import { formatDate } from "~/helpers/date";
 import { divideBigIntToFloat } from "~/helpers/numbers";
-import { useAccount } from "@/store/account.ts";
+import { useAccount } from "~/store/account.ts";
 
 const accountStore = useAccount();
