+++ components/overlays/OverlayDialog.vue
@@ -71,7 +71,7 @@ import {
   TransitionRoot,
 } from "@headlessui/vue";
 import { defineProps } from "vue";
-import TransitionChildSootGlass from "@/components/ui/TransitionChildSootGlass.vue";
+import TransitionChildSootGlass from "~/components/ui/TransitionChildSootGlass.vue";
 
 const props = defineProps({
   close: {
