--- pages/gov.vue
@@ -1,68 +1,0 @@
-<template>
-  <div class="mt-3 text-center sm:mt-5">
-    <span as="h3" class="text-base font-semibold leading-6 text-white"
-      >Private Voting
-    </span>
-
-    <div class="mt-2">
-      <p class="text-sm text-gray-400 text-justify">COMING SOON<sup>TM</sup></p>
-      <p class="text-sm text-gray-400 text-justify">
-        If you participate in OpenGov on Polkadot and Kusama or onchain
-        governance on any of their parachains, you inevitably expose your voting
-        preferences to the world. While your account can be a pseudonym, all
-        your actions will be linkable and eventually reveal your identity.
-        Knowing this will likely affect your voting behavior and promote
-        conformity, obedience and submission. Vote privacy, on the other hand,
-        rather promotes creativity, exploration and dissent.
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        Integritee will allow you to vote with enhanced privacy in web3
-        governance. This will enable you to vote freely and without fear of
-        repercussions.
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        With great power comes great responsibility. For web3 governance to be
-        legitimate, we need to strike a balance between privacy and
-        accountability. Dominant voting power should never come with absolute
-        privacy.
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        We encourage you to come
-        <a
-          href="https://forum.polkadot.network/t/private-voting-on-referenda/3001"
-          >discuss these topics with us.</a
-        >
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        We are currently working on the implementation of private voting. Stay
-        tuned for updates.
-      </p>
-      <!-- this is necessary to avoid the footer overlapping the text -->
-      <br /><br /><br /><br /><br /><br /><br />
-    </div>
-  </div>
-</template>
-
-<script setup lang="ts">
-// Import your components and add your setup logic here
-</script>
-
-<style scoped>
-h1 {
-  font-size: 2em; /* Adjust as needed */
-  font-weight: bold; /* Makes the text bold */
-  color: #999; /* Change color as needed */
-  text-align: center; /* Centers the text */
-  margin-bottom: 20px; /* Adds space below the heading */
-}
-
-p {
-  padding: 10px 20px; /* Adds 10px padding top and bottom, 20px padding left and right */
-  line-height: 1.5; /* Increases the space between lines of text */
-  color: #fff; /* Change color as needed */
-  text-align: justify;
-}
-a {
-  color: #99f; /* Change color as needed */
-}
-</style>
