--- pages/swap.vue
@@ -1,54 +1,0 @@
-<template>
-  <div class="mt-3 text-center sm:mt-5">
-    <span as="h3" class="text-base font-semibold leading-6 text-white"
-      >Private Swaps
-    </span>
-
-    <div class="mt-2">
-      <p class="text-sm text-gray-400 text-justify">COMING SOON<sup>TM</sup></p>
-      <p class="text-sm text-gray-400 text-justify">
-        Whenever you use a centralized exchange (CEX), you trust them to operate
-        a fair market and protect your data. Even worse, you hand over custody
-        over your assets to them. The upside is: fast trading and something like
-        privacy (in the best case your exposing your data and behavior to one
-        entity only, not to the public)
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        If you need more autonomy, you may want to use decentralized exchanges
-        (DEX). There, you keep custody over your assets and no single entity has
-        the power to deny you access to the trading platform. However, this
-        freedom comes with downsides: High latency and front-running
-      </p>
-      <p class="text-sm text-gray-400 text-justify">
-        Incognitee can support shielding of various digital assets and offer
-        automated market makers (AMM, similar to uniswap) on L2. This will
-        prevent front-running because no one can see the transactions before
-        they are executed. It also offers enhanced speed because of subsecond
-        block times on Incognitee
-      </p>
-      <!-- this is necessary to avoid the footer overlapping the text -->
-      <br /><br /><br /><br /><br /><br /><br />
-    </div>
-  </div>
-</template>
-
-<script setup lang="ts">
-// Import your components and add your setup logic here
-</script>
-
-<style scoped>
-h1 {
-  font-size: 2em; /* Adjust as needed */
-  font-weight: bold; /* Makes the text bold */
-  color: #999; /* Change color as needed */
-  text-align: center; /* Centers the text */
-  margin-bottom: 20px; /* Adds space below the heading */
-}
-
-p {
-  padding: 10px 20px; /* Adds 10px padding top and bottom, 20px padding left and right */
-  line-height: 1.5; /* Increases the space between lines of text */
-  color: #fff; /* Change color as needed */
-  text-align: justify;
-}
-</style>
